requirements.txt
ansible==2.6.4
asn1crypto==0.24.0
astroid==2.0.4
attrs==17.4.0
-e git+git@code.greenhost.net:greenhost/barbabot.git@7281d2f5f09883fee4607fdb8de9b56ad6d9a9e0#egg=barbabot
bcrypt==3.1.4
cffi==1.11.5
cryptography==2.3.1
Django==2.1.3
djangorestframework==3.9.0
idna==2.7
install-requires==0.3.0
isort==4.3.4
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mypy==0.560
paramiko==2.4.2
pluggy==0.6.0
psutil==5.4.3
py==1.5.2
pyasn1==0.4.4
pycodestyle==2.3.1
pycparser==2.19
pydocstyle==2.1.1
pylint==2.1.1
pylint-django==2.0.2
pylint-plugin-utils==0.4
PyNaCl==1.3.0
pytest==3.4.1
pytz==2018.7
PyYAML==3.13
six==1.11.0
snowballstemmer==1.2.1
typed-ast==1.1.0
virtualenv==15.1.0
wrapt==1.10.11
